pyproject.toml
[project]
name = "typinggame"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "json5>=0.12.0",
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "pillow>=11.2.1",
    "pydub>=0.25.1",
    "pygame>=2.6.1",
    "pyopengl>=3.1.9",
    "pyopengl-accelerate>=3.1.9",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "seaborn>=0.13.2",
    "opencv-python>=4.10.0",
    "scikit-image>=0.24.0",
]
